requirements.txt
altgraph==0.17.4
asgiref==3.7.1
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
black==23.3.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
cssselect2==0.7.0
cx-Freeze==6.15.10
cx-Logging==3.1.0
decorator==5.1.1
Django==4.2.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fonttools==4.43.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.13.2
jedi==0.18.2
kaitaistruct==0.10
lief==0.13.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
numpy==1.24.3
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathspec==0.11.1
pdfkit==1.0.0
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.0.1
pipdeptree==2.13.0
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
py2exe==0.13.0.1
pyasn1==0.5.0
pycparser==2.21
pydivert==2.1.0
pydyf==0.8.0
Pygments==2.15.1
pyinstaller-hooks-contrib==2023.10
pyOpenSSL==23.2.0
pyparsing==3.1.0
pyphen==0.14.0
PySocks==1.7.1
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.2
requests==2.31.0
selenium==4.13.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.2
tinycss2==1.2.1
tk==0.1.0
tomli==2.0.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
weasyprint==60.1
webencodings==0.5.1
wsproto==1.2.0
zopfli==0.2.3
zstandard==0.21.0
